main,카테고리

+++ oneday_class/src/main/java/kr/co/oneclass/search/CreatorDTO.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Setter
 @ToString
 public class CreatorDTO {
-    private long operatorCode;
-    private long memberCode;
+    private int operatorCode;
+    private int memberCode;
     private String approvalStatus;
     private String activityName;        // 작가/활동명
     private String profileImage;
